compile on linux

+++ inc/util.hpp
@@ -5,7 +5,7 @@
 
 //std::string executeCgi(const Request &request, const std::string &cgiBinPath, std::string file_content);
 
-ssize_t stoi(std::string content_length);
+int stoi(std::string content_length);
 std::string itos(int n);
 
 #endif
